pyproject.toml
[project]
name = "sales-call-insights-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai-haystack>=2.2.0",
    "haystack-ai>=2.20.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]
